+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## UPCOMING
 
+* fix export image highcharts
 * fix import bug with format SERPOSCOPE_CSV
 
 ## 1.0.4 - 21/06/2013