+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Serposcope changelog
 
-## UPCOMING
+## 1.0.5 - 05/09/2013
 
+* fix SSL redirection introduced by Google in september
 * fix export image highcharts
 * fix import bug with format SERPOSCOPE_CSV
 * run cron for a specific group `/usr/bin/php cron.php _ID_GROUP_